// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Tone = {
  PROFESSIONAL: 'PROFESSIONAL',
  CASUAL: 'CASUAL',
  EXCITING: 'EXCITING',
  TRUSTWORTHY: 'TRUSTWORTHY'
} as const

export type Tone = (typeof Tone)[keyof typeof Tone]


export const GenerationStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type GenerationStatus = (typeof GenerationStatus)[keyof typeof GenerationStatus]